Consolidate the two TaskStatus classes.
There are two, but we don't need more than one.
