+++ core/src/main/java/org/apache/druid/indexer/TaskStatus.java
@@ -96,11 +96,11 @@ public class TaskStatus
   private final TaskLocation location;
 
   @JsonCreator
-  protected TaskStatus(
+  public TaskStatus(
       @JsonProperty("id") String id,
       @JsonProperty("status") TaskState status,
       @JsonProperty("duration") long duration,
-      @JsonProperty("errorMsg") @Nullable String errorMsg,
+      @Nullable @JsonProperty("errorMsg") String errorMsg,
       @Nullable @JsonProperty("location") TaskLocation location
   )
   {
