+++ indexing-service/src/main/java/org/apache/druid/indexing/worker/shuffle/LocalIntermediaryDataManager.java
@@ -25,7 +25,7 @@ import com.google.common.io.ByteSource;
 import com.google.common.io.Files;
 import com.google.inject.Inject;
 import org.apache.commons.lang3.mutable.MutableInt;
-import org.apache.druid.client.indexing.TaskStatus;
+import org.apache.druid.indexer.TaskStatus;
 import org.apache.druid.common.guava.FutureUtils;
 import org.apache.druid.common.utils.IdUtils;
 import org.apache.druid.guice.ManageLifecycle;
