--- server/src/main/java/org/apache/druid/client/indexing/TaskStatus.java
@@ -1,101 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one
- * or more contributor license agreements.  See the NOTICE file
- * distributed with this work for additional information
- * regarding copyright ownership.  The ASF licenses this file
- * to you under the Apache License, Version 2.0 (the
- * "License"); you may not use this file except in compliance
- * with the License.  You may obtain a copy of the License at
- *
- *   http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied.  See the License for the
- * specific language governing permissions and limitations
- * under the License.
- */
-
-package org.apache.druid.client.indexing;
-
-import com.fasterxml.jackson.annotation.JsonCreator;
-import com.fasterxml.jackson.annotation.JsonProperty;
-import com.google.common.base.Objects;
-import com.google.common.base.Preconditions;
-import org.apache.druid.indexer.TaskState;
-
-/**
- * Should be synced with org.apache.druid.indexing.common.TaskStatus
- */
-public class TaskStatus
-{
-  private final String id;
-  private final TaskState status;
-  private final long duration;
-
-  @JsonCreator
-  public TaskStatus(
-      @JsonProperty("id") String id,
-      @JsonProperty("status") TaskState status,
-      @JsonProperty("duration") long duration
-  )
-  {
-    this.id = id;
-    this.status = status;
-    this.duration = duration;
-
-    // Check class invariants.
-    Preconditions.checkNotNull(id, "id");
-    Preconditions.checkNotNull(status, "status");
-  }
-
-  @JsonProperty("id")
-  public String getId()
-  {
-    return id;
-  }
-
-  @JsonProperty("status")
-  public TaskState getStatusCode()
-  {
-    return status;
-  }
-
-  @JsonProperty("duration")
-  public long getDuration()
-  {
-    return duration;
-  }
-
-  @Override
-  public boolean equals(Object o)
-  {
-    if (this == o) {
-      return true;
-    }
-    if (o == null || getClass() != o.getClass()) {
-      return false;
-    }
-    TaskStatus that = (TaskStatus) o;
-    return duration == that.duration &&
-           java.util.Objects.equals(id, that.id) &&
-           status == that.status;
-  }
-
-  @Override
-  public int hashCode()
-  {
-    return java.util.Objects.hash(id, status, duration);
-  }
-
-  @Override
-  public String toString()
-  {
-    return Objects.toStringHelper(this)
-                  .add("id", id)
-                  .add("status", status)
-                  .add("duration", duration)
-                  .toString();
-  }
-}
