+++ server/src/main/java/org/apache/druid/rpc/indexing/OverlordClient.java
@@ -20,8 +20,8 @@
 package org.apache.druid.rpc.indexing;
 
 import com.google.common.util.concurrent.ListenableFuture;
-import org.apache.druid.client.indexing.TaskStatus;
 import org.apache.druid.client.indexing.TaskStatusResponse;
+import org.apache.druid.indexer.TaskStatus;
 import org.apache.druid.rpc.ServiceRetryPolicy;
 
 import java.util.Map;
